// medical-clinics-frontend/src/app/services/models/new-note-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface NewNoteRequest {
  chiefComplaint?: string;
  diagnosis: string;
  diagnosisCode?: string;
  examination: string;
  medicalHistory: string;
  medicationAndRecommendations?: string;
  testResults?: string;
}